+++ matcher/defualts.go
@@ -6,7 +6,7 @@ import (
 )
 
 // Points to the distance/LevenshteinDistance function
-func DefualtDistance(query, target string) int32 {
+func DefualtDistance(query, target string) float64 {
 	return distance.LevenshteinDistance(query, target)
 }
 
@@ -16,7 +16,7 @@ func DefualtString(input string) string {
 }
 
 // Linear function for the score
-func DefualtScore(query, target string, dist int32) float64 {
+func DefualtScore(query, target string, dist float64) float64 {
 	return score.Linear(query, target, dist)
 }
