+++ suddig.go
@@ -26,7 +26,7 @@ func Score(query, target string) float64 {
 // The Levenshtein algorithm returns a number based on how many edits is needed for the query to match the target.
 // O(m*n) where m & n are the length of the input strigns.
 // See Score for the percental difference between the strings.
-func Distance(query, target string) int32 {
+func Distance(query, target string) float64 {
 	m := matcher.New(matcher.DefualtConfig())
 	return m.Distance(query, target)
 }
